Change the formula to calculate the WQI
What: the previous formula was incomplete. It only worked for the cases
where the measure increments (then the WQI increments) but it didn't work
for any other case. So when the pH or the DO decrements the formula decrements
the WQI, which is not correct since a lower WQI means better water quality.